Move block range and execution plan into dedicated files
