+++ gossip/blockproc/bundle/bundle_test.go
@@ -19,14 +19,12 @@ package bundle
 import (
 	"errors"
 	"fmt"
-	"math"
 	"math/big"
 	"testing"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/ethereum/go-ethereum/rlp"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
 )
@@ -178,268 +176,6 @@ func TestExtractExecutionPlan_FailsIfPlanExtractionFails(t *testing.T) {
 
 	_, got := ExtractExecutionPlan(signer, envelope)
 	require.Equal(want, got)
-}
-
-func TestMakeMaxRangeStartingAt_CreatesMaxRangeStartingAtGivenBlock(t *testing.T) {
-	cases := map[string]struct {
-		start          uint64
-		expectedLatest uint64
-		expectedSize   uint64
-	}{
-		"start at 0": {
-			start:          0,
-			expectedLatest: MaxBlockRange - 1,
-			expectedSize:   MaxBlockRange,
-		},
-		"start at 1": {
-			start:          1,
-			expectedLatest: MaxBlockRange,
-			expectedSize:   MaxBlockRange,
-		},
-		"start at 100": {
-			start:          100,
-			expectedLatest: 100 + MaxBlockRange - 1,
-			expectedSize:   MaxBlockRange,
-		},
-		"start with max plus one blocks": {
-			start:          math.MaxUint64 - MaxBlockRange - 1,
-			expectedLatest: math.MaxUint64 - 2,
-			expectedSize:   MaxBlockRange,
-		},
-		"start with max blocks": {
-			start:          math.MaxUint64 - MaxBlockRange,
-			expectedLatest: math.MaxUint64 - 1,
-			expectedSize:   MaxBlockRange,
-		},
-		"start with exact left blocks": {
-			start:          math.MaxUint64 - MaxBlockRange + 1,
-			expectedLatest: math.MaxUint64,
-			expectedSize:   MaxBlockRange,
-		},
-		"start with not enough blocks": {
-			start:          math.MaxUint64 - MaxBlockRange + 2,
-			expectedLatest: math.MaxUint64,
-			expectedSize:   MaxBlockRange - 1,
-		},
-		"start with two blocks left": {
-			start:          math.MaxUint64 - 1,
-			expectedLatest: math.MaxUint64,
-			expectedSize:   2,
-		},
-		"start with one block left": {
-			start:          math.MaxUint64,
-			expectedLatest: math.MaxUint64,
-			expectedSize:   1,
-		},
-	}
-	for name, c := range cases {
-		t.Run(name, func(t *testing.T) {
-			r := MakeMaxRangeStartingAt(c.start)
-			require.Equal(t, c.start, r.Earliest)
-			require.Equal(t, c.expectedLatest, r.Latest)
-			require.Equal(t, c.expectedSize, r.Size())
-		})
-	}
-}
-
-func TestBlockRange_Size_ReturnsCorrectSize(t *testing.T) {
-	tests := map[string]struct {
-		blockRange BlockRange
-		want       uint64
-	}{
-		"empty range": {
-			blockRange: BlockRange{
-				Earliest: 10,
-				Latest:   9,
-			},
-			want: 0,
-		},
-		"single range": {
-			blockRange: BlockRange{
-				Earliest: 10,
-				Latest:   10,
-			},
-			want: 1,
-		},
-		"two blocks range": {
-			blockRange: BlockRange{
-				Earliest: 10,
-				Latest:   11,
-			},
-			want: 2,
-		},
-		"multiple blocks range": {
-			blockRange: BlockRange{
-				Earliest: 10,
-				Latest:   20,
-			},
-			want: 11,
-		},
-		"large range": {
-			blockRange: BlockRange{
-				Earliest: 0,
-				Latest:   10_000_000,
-			},
-			want: 10_000_001,
-		},
-		"large range with latest near max uint64": {
-			blockRange: BlockRange{
-				Earliest: 0,
-				Latest:   math.MaxUint64 - 1,
-			},
-			want: math.MaxUint64,
-		},
-		"too large range is capped to prevent overflow": {
-			blockRange: BlockRange{
-				Earliest: 0,
-				Latest:   math.MaxUint64,
-			},
-			want: math.MaxUint64,
-		},
-		"small range start near max uint64": {
-			blockRange: BlockRange{
-				Earliest: math.MaxUint64 - 10,
-				Latest:   math.MaxUint64,
-			},
-			want: 11,
-		},
-		"small range with the last two blocks": {
-			blockRange: BlockRange{
-				Earliest: math.MaxUint64 - 1,
-				Latest:   math.MaxUint64,
-			},
-			want: 2,
-		},
-		"single block range at max uint64": {
-			blockRange: BlockRange{
-				Earliest: math.MaxUint64,
-				Latest:   math.MaxUint64,
-			},
-			want: 1,
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			require.EqualValues(t, test.want, test.blockRange.Size())
-		})
-	}
-}
-
-func TestBlockRange_IsInRange_ReturnsTrueIfBlockNumberIsWithinRange(t *testing.T) {
-	tests := map[string]struct {
-		BlockRange BlockRange
-		current    uint64
-		want       bool
-	}{
-		"within range": {
-			BlockRange: BlockRange{Earliest: 10, Latest: 20},
-			current:    15,
-			want:       true,
-		},
-		"at earliest": {
-			BlockRange: BlockRange{Earliest: 10, Latest: 20},
-			current:    10,
-			want:       true,
-		},
-		"at latest": {
-			BlockRange: BlockRange{Earliest: 10, Latest: 20},
-			current:    20,
-			want:       true,
-		},
-		"below range": {
-			BlockRange: BlockRange{Earliest: 10, Latest: 20},
-			current:    9,
-			want:       false,
-		},
-		"above range": {
-			BlockRange: BlockRange{Earliest: 10, Latest: 20},
-			current:    21,
-			want:       false,
-		},
-		"at lower end": {
-			BlockRange: BlockRange{Earliest: 10, Latest: 20},
-			current:    10,
-			want:       true,
-		},
-		"at upper end": {
-			BlockRange: BlockRange{Earliest: 10, Latest: 20},
-			current:    20,
-			want:       true,
-		},
-		"single block range": {
-			BlockRange: BlockRange{Earliest: 10, Latest: 10},
-			current:    10,
-			want:       true,
-		},
-		"invalid range": {
-			BlockRange: BlockRange{Earliest: 20, Latest: 10},
-			current:    15,
-			want:       false,
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			got := test.BlockRange.IsInRange(test.current)
-			require.Equal(t, test.want, got)
-		})
-	}
-}
-
-func TestExecutionPlan_Hash_ComputesDeterministicHash(t *testing.T) {
-
-	step1 := ExecutionStep{
-		From: common.HexToAddress("0x0000000000000000000000000000000000000001"),
-		Hash: common.Hash{0x01},
-	}
-	step2 := ExecutionStep{
-		From: common.HexToAddress("0x0000000000000000000000000000000000000002"),
-		Hash: common.Hash{0x02},
-	}
-
-	tests := map[string]ExecutionPlan{
-		"empty plan": {},
-		"plan with transactions": {
-			Steps: []ExecutionStep{step1, step2},
-		},
-		"plan with flag 1": {
-			Steps: []ExecutionStep{step1},
-			Flags: 0x1,
-		},
-		"plan with flag 2": {
-			Steps: []ExecutionStep{step1},
-			Flags: 0x2,
-		},
-		"plan with flag 3": {
-			Steps: []ExecutionStep{step1},
-			Flags: 0x3,
-		},
-	}
-
-	seenHashes := make(map[common.Hash]struct{})
-	for name, executionPlan := range tests {
-		t.Run(name, func(t *testing.T) {
-
-			transactions := make([]any, len(executionPlan.Steps))
-			for i, step := range executionPlan.Steps {
-				transactions[i] = []any{step.From, step.Hash}
-			}
-			manualSerialize := []any{
-				transactions,
-				executionPlan.Flags,
-				executionPlan.Range,
-			}
-
-			hasher := crypto.NewKeccakState()
-			require.NoError(t, rlp.Encode(hasher, manualSerialize))
-			computed := common.BytesToHash(hasher.Sum(nil))
-
-			require.Equal(t, executionPlan.Hash(), computed)
-			require.NotContains(t, seenHashes, computed, "hash should be unique for different plans")
-			seenHashes[computed] = struct{}{}
-		})
-	}
 }
 
 func TestExtractExecutionPlan_ExtractsStepsAndFlags(t *testing.T) {
